requirements.txt
anyascii==0.3.2
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.1
defusedxml==0.7.1
dj-database-url==2.3.0
Django==5.1.4
django-cogwheels==0.3
django-filter==24.3
django-modelcluster==6.4
django-permissionedforms==0.1
django-taggit==6.1.0
django-treebeard==4.7.1
djangorestframework==3.15.2
draftjs-exporter==5.0.0
et_xmlfile==2.0.0
filetype==1.2.0
idna==3.10
l18n==2021.3
laces==0.1.1
openpyxl==3.1.5
pillow==11.0.0
pillow_heif==0.21.0
psycopg==3.2.4
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
telepath==0.3.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
wagtail==6.3.1
wagtail-modeladmin==2.1.0
wagtailmenus==4.0.2
Willow==1.9.0
